src/requirements/base.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
python-decouple==3.8
drf-spectacular==0.28.0
django-uuslug==2.0.0
celery==5.4.0
ruff==0.6.9